backend/requirements.txt
# to install from app -  pip install -r ../requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
Flask-Cors
Flask-JWT-Extended
Flask-Limiter
pydantic
email-validator
bcrypt
gunicorn



boto3
requests